+++ common-library/src/main/java/com/library/common/annotation/RequiresPermission.java
@@ -1,0 +1,12 @@
+package com.library.common.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.CLASS)
+public @interface RequiresPermission {
+
+}
