+++ common-library/src/main/java/com/library/common/annotation/ValidateMember.java
@@ -1,0 +1,11 @@
+package com.library.common.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ValidateMember {
+}
