+++ common-library/src/main/java/com/library/common/dto/MemberInfoDto.java
@@ -1,0 +1,16 @@
+package com.library.common.dto;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class MemberInfoDto {
+
+    private Long memberId;
+    private String name;
+
+}
