+++ common-library/src/main/resources/application.yml
@@ -1,0 +1,12 @@
+spring:
+  application:
+    name: validate-library
+  data:
+    redis:
+      host: localhost
+      port: 6379
+eureka:
+  client:
+    register-with-eureka: false
+
+
