+++ common-library/src/test/java/com/library/common/CommonLibraryApplicationTests.java
@@ -1,0 +1,13 @@
+package com.library.common;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CommonLibraryApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
