+++ member-service/src/main/java/com/prs/ms/MemberServiceApplication.java
@@ -3,7 +3,7 @@ package com.prs.ms;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication(scanBasePackages = {"com.library.validate.handlers", "com.prs.ms"})
+@SpringBootApplication(scanBasePackages = {"com.library.common.handlers", "com.prs.ms"})
 public class MemberServiceApplication {
 
     public static void main(String[] args) {
