+++ platform-service/build.gradle
@@ -59,7 +59,7 @@ dependencies {
 	annotationProcessor "jakarta.annotation:jakarta.annotation-api"
 	annotationProcessor "jakarta.persistence:jakarta.persistence-api"
 
-	implementation 'com.library:validate:1.0'
+	implementation 'com.library:common:1.0'
 
 
 }
