+++ platform-service/src/main/java/com/prs/ps/PlatformServiceApplication.java
@@ -2,6 +2,7 @@ package com.prs.ps;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.ComponentScans;
@@ -10,7 +11,8 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @SpringBootApplication
 @EnableJpaAuditing
 @EnableFeignClients
-@ComponentScan(basePackages = {"com.library.validate", "com.prs.ps"})
+@EnableCaching
+@ComponentScan(basePackages = {"com.library.common", "com.prs.ps"})
 public class PlatformServiceApplication {
 
     public static void main(String[] args) {
