+++ review-service/build.gradle
@@ -37,7 +37,7 @@ dependencies {
 	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
 	implementation 'org.springframework.kafka:spring-kafka'
 
-	implementation 'com.library:validate:1.0'
+	implementation 'com.library:common:1.0'
 
 	compileOnly 'org.projectlombok:lombok'
 	runtimeOnly 'com.mysql:mysql-connector-j'
