+++ review-service/src/main/java/com/prs/rs/aop/PlatformValidateAspect.java
@@ -1,6 +1,6 @@
 package com.prs.rs.aop;
 
-import static com.library.validate.utils.CommonAopUtils.*;
+import static com.library.common.utils.CommonAopUtils.*;
 
 import com.prs.rs.annotation.ValidatePlatform;
 import com.prs.rs.client.PlatformServiceClient;
