+++ review-service/src/main/java/com/prs/rs/controller/ReviewController.java
@@ -1,6 +1,6 @@
 package com.prs.rs.controller;
 
-import com.library.validate.dto.MemberInfoDto;
+import com.library.common.dto.MemberInfoDto;
 import com.prs.rs.domain.Review;
 import com.prs.rs.dto.request.ReviewEditDto;
 import com.prs.rs.dto.request.ReviewListDto;
