--- validate-library/src/main/java/com/library/validate/annotation/RequiresPermission.java
@@ -1,12 +1,0 @@
-package com.library.validate.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.CLASS)
-public @interface RequiresPermission {
-
-}
