--- validate-library/src/main/java/com/library/validate/annotation/ValidateMember.java
@@ -1,11 +1,0 @@
-package com.library.validate.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ValidateMember {
-}
