--- validate-library/src/main/java/com/library/validate/config/FeignConfig.java
@@ -1,16 +1,0 @@
-package com.library.validate.config;
-
-import com.library.validate.client.FeignCookieInterceptor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class FeignConfig {
-
-
-    @Bean
-    public FeignCookieInterceptor feignCookieInterceptor() {
-        return new FeignCookieInterceptor();
-    }
-
-}
