--- validate-library/src/main/java/com/library/validate/dto/MemberInfoDto.java
@@ -1,16 +1,0 @@
-package com.library.validate.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class MemberInfoDto {
-
-    private Long memberId;
-    private String name;
-
-}
