--- validate-library/src/test/java/com/library/validate/ValidateLibraryApplicationTests.java
@@ -1,13 +1,0 @@
-package com.library.validate;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ValidateLibraryApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
